Remove --enable-heuristic/--dfs, and add --good-enough
The idea is, `ap`, by default, would always try to look for the optimal
solution; however, depending on the input, searching for the optimal
schedule might cause `ap` to crash, or to spin forever.  The solution?
Simply try to settle for a sub-optimal solution (it's hard to judge the
quality of it, given that we don't have anything to compare with, but
still better than nothing).

So...

- Optimal solution: A* without heuristic
- Good enough solution: it will take the best between

  - A* with very naive solution (one that can easily overestimate the
  remaining cost)
  - Bounded DFS -- i.e. it will evaluate a maximum of 50000 sates before
  returning the best